requirements.txt
numpy<1.24.0
fire==0.1.3
soundfile==0.10.3.post1
unidecode==1.1.1
numba==0.48.0
librosa==0.7.2
mpi4py>=3.0.0
av==8
git+https://github.com/jmoso13/a-unet.git@add_dilation
git+https://github.com/jmoso13/audio-diffusion-pytorch.git@noise_pass
pytorch-lightning==1.7
wandb
Pygments==2.14.0
git+https://github.com/arogozhnikov/einops.git@master
ipython
pandas
resampy==0.3.1
tqdm==4.57.0
pillow
git+https://github.com/jmoso13/jukebox.git
PyYAML
git+https://github.com/Harmonai-org/sample-generator.git
git+https://github.com/crowsonkb/v-diffusion-pytorch.git